+++ craps-simulator/Interfaces/IBetResult.cs
@@ -1,8 +1,8 @@
 ﻿namespace craps_simulator.Interfaces {
     public interface IBetResult {
-        float Bet { get; set; }
+        decimal Bet { get; set; }
         bool IsWinner { get; set; }
         bool IsLoser { get; set; }
-        float Pays { get; set; }
+        decimal Pays { get; set; }
     }
 }
